task1/requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
web3==5.23.0
pymongo==3.12.0
requests==2.25.1
python-dotenv
pymongo[srv]
